[fix:objective-functions-losses] refactor: objective unit tests according to the new file structure
